fixed clicking bug and added ProgressBar component
